requirement.txt
easydict==1.9
numpy
pandas==1.1.5
git+git://github.com/erikwijmans/Pointnet2_PyTorch.git#egg=pointnet2_ops&subdirectory=pointnet2_ops_lib
pomegranate
pyquaternion==0.9.9
pytorch-lightning==1.3.8
PyYAML==5.4.1
scipy==1.5.4
Shapely==1.7.1
tdqm==0.0.1
tensorboard==2.4.1
torchmetrics==0.4.1
tqdm==4.61.1
transforms3d
learn2learn
chamferdist
hausdorff
